+++ .vscode/settings.json
@@ -30,6 +30,7 @@
         "COUNTA",
         "CRDT",
         "customd",
+        "dashedrect",
         "Datas",
         "DATEFMTLISG",
         "demi",
@@ -106,6 +107,7 @@
         "rbush",
         "redi",
         "rowheader",
+        "scrolltrack",
         "shrinkability",
         "Sider",
         "strechability",
